Git Repo Work
Refactored the GitContentRepository class into three classes to help reduce the size. Basic boostrapping and site creation from blueprint are working at this point.
