+++ src/main/resources/crafter/studio/studio-services-context.xml
@@ -110,7 +110,6 @@
         <property name="importService" ref="cstudioImportService"/>
         <property name="generalLockService" ref="cstudioGeneralLockService"/>
         <property name="rebuildRepositoryMetadata" ref="studioRebuildRepositoryMetadata"/>
-        <property name="createSiteV2" value="${siteService.createSite.v2.enable}"/>
         <property name="studioConfiguration" ref="studioConfiguration"/>
     </bean>
